# Conflicts:
#	Programs/QuickEd/Classes/UI/Preview/PreviewWidget.cpp
#	Programs/QuickEd/Classes/UI/Preview/PreviewWidget.h


Former-commit-id: 6518e345e30bf3a107e106c7a0676c731b1a9de9

+++ Sources/Internal/Scene3D/DataNode.h
@@ -16,7 +16,7 @@ namespace DAVA
 class DataNode : public BaseObject
 {
 public:
-    static const int32 INVALID_ID = 0;
+    static const uint64 INVALID_ID = 0;
 
 public:
     DataNode();
